feat: Storage Layer Refactoring Phase 3B - Price Domain Extraction (#121)
This commit extracts all price-related database operations from the monolithic
server/storage.ts into a dedicated PriceStorage domain repository, following
the established patterns from Phase 1 (Foundation), Phase 2 (UserStorage), and
Phase 3A (ProductStorage).

## What Changed

### New Files
- **server/storage/domains/price-storage.ts** (~1,100 lines)
  - Extracted 28 price-related methods from storage.ts
  - Organized into 5 logical sections for maintainability
  - Implements all 7-point implementation guidance from base-storage.ts

### Modified Files
- **server/storage.ts**
  - Added PriceStorage property and initialization
  - Converted 28 methods to trivial delegation pattern
  - Maintains backward compatibility with existing IStorage interface

## Methods Extracted (28 total)

### 1. Price History Operations (4 methods)
- getPriceHistory() - Smart data source selection (raw/daily/weekly/monthly)
- getRetailerPriceHistory() - Retailer-specific price history
- getPriceHistoryByOfferId() - Offer-specific history for drop detection
- getLatestPriceForOffer() - Most recent price for an offer

### 2. Price Trend & Analysis (2 methods)
- analyzePriceTrend() - 30-day trend analysis with statistics
- getBestTimeToBuy() - Purchase timing recommendations

### 3. Price Analytics & Aggregation (9 methods)
- getWeeklyAggregates() - ISO week-based aggregations
- getMonthlyAggregates() - Calendar month aggregations
- getDailyAggregates() - Daily price statistics
- insertPriceHistory() - Raw price record insertion
- upsertDailyAggregate() - Daily aggregate upsert
- upsertWeeklyAggregate() - Weekly aggregate upsert
- upsertMonthlyAggregate() - Monthly aggregate upsert
- getWeeklyAggregatesByRange() - Multi-week batch query
- getMonthlyAggregatesByRange() - Multi-month batch query

### 4. Price Snapshot Operations (8 methods)
- createPriceSnapshot() - Daily price snapshot creation
- getPriceSnapshots() - Historical snapshot retrieval
- getLatestPriceSnapshot() - Most recent snapshot
- queryPriceHistory() - Advanced filtered queries
- insertBulkPriceHistory() - Batch price insertion (500 records/chunk)
- getProductOffersForSnapshot() - Batch offer retrieval
- getAggregationData() - Raw data for aggregation jobs
- deleteStalePriceHistory() - 90-day retention cleanup

### 5. Trend Analysis Operations (5 methods)
- upsertPriceTrends() - Batch trend updates (100 records/chunk)
- getPriceTrends() - Product trend retrieval with retailer info
- getAnalyticsOverview() - Dashboard statistics
- getTrendPriceData() - 30-day trend calculation data
- getPriceSnapshotsByDateRange() - Snapshot queries

## Implementation Quality

### Input Validation (✓ Implemented)
- validateProductId() - Positive integer validation for product IDs
- validateOfferId() - Positive integer validation for offer IDs
- validateRetailerId() - Positive integer validation for retailer IDs
- All numeric parameters validated before database operations

### N+1 Query Prevention (✓ Implemented)
- No queries in loops - all use JOINs or batch operations
- Uses array_agg() for grouped price data (getAggregationData)
- Uses inArray() for batch offer retrieval (getProductOffersForSnapshot)
- LEFT JOINs with retailers table for name enrichment

### Database Aggregation (✓ Optimized)
- Server-side aggregations reduce memory usage
- Smart data source selection (raw → daily → weekly → monthly)
- Batch operations chunked to avoid PostgreSQL parameter limits

### Error Handling (✓ Consistent)
- All methods use handleError() from BaseStorage
- Validation errors include context (e.g., "Invalid productId: -1")
- Throws immediately on validation failures

### Type Safety (✓ Enforced)
- No inline types in method signatures
- All types from storage/types.ts or @shared/schema
- Proper NormalizedPricePoint type for price-history-service integration

## Statistics

- **Lines Added**: ~1,100 (price-storage.ts)
- **Lines Modified**: ~60 (storage.ts delegation updates)
- **Methods Extracted**: 28
- **Validation Helpers**: 3 (validateProductId, validateOfferId, validateRetailerId)
- **Import Path Fixes**: 2 (price-history-service dynamic imports)
- **TypeScript Errors**: 0 NEW (pre-existing errors unchanged)

## Testing

### TypeScript Compilation
```bash
npm run check
# Result: Zero NEW errors introduced
```

### Code Review
- ✓ Reviewed by code-review-specialist agent
- ✓ All 7-point implementation guidance followed
- ✓ Type safety verified (NormalizedPricePoint)
- ✓ N+1 prevention patterns validated
- ✓ Input validation coverage confirmed

## Migration Impact

### Breaking Changes
- None - All changes are internal refactoring
- IStorage interface unchanged
- Method signatures preserved
- Backward compatibility maintained

### Database Impact
- No schema changes
- No migrations required
- Query performance unchanged

### Route Impact
- No route changes required
- All routes continue using storage.getX() pattern
- Service layer unaffected

## Related Issues

- Addresses #121 (Storage Layer Refactoring)
- Follows patterns from Phase 3A completion (#139)
- Part of larger storage domain extraction initiative

## Next Steps (Phase 3C+)

Remaining domains for extraction:
- Forum operations (~15 methods)
- Retailer operations (~8 methods)
- Admin analytics (~10 methods)
- Notification system (~6 methods)
- Alert management (~5 methods)
- Watch list operations (~12 methods)
- Community features (~8 methods)
